Refactor enrich into submodules
